fix: Use a DocumentFragment to build cat list
Forgot the var when declaring the catListElement variable. When I added
it that created a closure problem. Used 'this' to fix that problem, but
that created another closure problem in my forEach loop. I was not able
to access catListElement inside the loop to append each li to it.

Solved that by using a DocumentFragment in the forEach loop to append
each li to. Then I was able to append the completed DocumentFragment to
the catListElement outside of the loop. Using a DocumentFragment is
better for performance too since it is not part of the DOM.